refactor: split up query classes for easier testing
